+++ apps/src/sites/studio/pages/pd/workshops/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import WorkshopMarketingPage from '@cdo/apps/code-studio/pd/workshop/WorkshopMarketingPage';
+import WorkshopMarketingPage from '@cdo/apps/code-studio/pd/workshops/WorkshopMarketingPage';
 import getScriptData from '@cdo/apps/util/getScriptData';
 
 document.addEventListener('DOMContentLoaded', function () {
